made a storage class
